split out sinks and sources
