+++ src/sink/index.ts
@@ -1,0 +1,3 @@
+// sink/index.ts
+
+export { default as toArray } from './to-array';
