+++ src/source/index.ts
@@ -1,0 +1,6 @@
+// source/index.ts
+
+export { default as all } from './all';
+export { default as from } from './from';
+export { default as merge } from './merge';
+export { default as series } from './series';
